Fix ruff lint errors and mypy type error in gem_scanner
Agent-Logs-Url: https://github.com/kishore446/360-Crypto-scalping-V2/sessions/85d28b43-5c87-45ed-9b75-97e7f355d92e

+++ src/gem_scanner.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 import time
 from dataclasses import dataclass, field
 from datetime import date
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Dict, Optional, Tuple
 
 import numpy as np
 
@@ -175,7 +175,7 @@ class GemScanner:
         if avg_vol <= 0:
             return None
 
-        vol_ratio = recent_vol / avg_vol
+        vol_ratio = float(recent_vol / avg_vol)
 
         if vol_ratio < self._config.min_volume_ratio:
             return None
